// lib/types.ts
export type StudentLevel = "A1" | "A2" | "B1" | "B2" | "C1" | "C2";
export type LessonStatus =
  | "Scheduled"
  | "Completed"
  | "Canceled"
  | "Rescheduled"
  | "Student no-show"
  | "Teacher no-show";
export type LessonMode = "Online" | "In person";
export type PaymentStatus = "Paid" | "Pending" | "Overdue";